added hooks and websockets
