+++ generators/app/templates/package.json
@@ -4,10 +4,11 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.12",
     "@fortawesome/free-solid-svg-icons": "^5.6.3",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "@inrupt/solid-style-guide": "^0.1.21",
     "@inrupt/solid-react-components": "0.3.0",
+    "@inrupt/solid-style-guide": "^0.1.21",
     "@rdfjs/data-model": "^1.1.1",
     "@solid/query-ldflex": "^2.4.0",
+    "@solid/react": "^1.4.34",
     "@svgr/webpack": "2.4.1",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "9.0.0",
@@ -50,13 +51,14 @@
     "postcss-loader": "3.0.0",
     "postcss-preset-env": "6.3.1",
     "postcss-safe-parser": "4.0.1",
-    "react": "^16.8.2",
+    "react": "^16.8.4",
     "react-app-polyfill": "^0.2.0",
     "react-click-outside": "^3.0.1",
     "react-dev-utils": "^7.0.1",
-    "react-dom": "^16.8.2",
+    "react-dom": "^16.8.4",
     "react-i18next": "^10.1.2",
     "react-router-dom": "^4.3.1",
+    "react-scripts": "^2.1.8",
     "react-toast-notifications": "^1.3.1",
     "resolve": "1.8.1",
     "sass-loader": "7.1.0",
@@ -84,11 +86,6 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "husky": {
-    "hooks": {
-      "pre-commit": "jest && eslint --ignore-path .gitignore ."
-    }
-  },
   "jest": {
     "collectCoverageFrom": [
       "src/**/*.{js,jsx,ts,tsx}",
@@ -148,7 +145,6 @@
             "@services": "./src/services",
             "@layouts": "./src/layouts",
             "@contexts": "./src/contexts/",
-            "@testSetup": "./src/utils/enzymeSetup",
             "@hocs": "./src/hocs",
             "@utils": "./src/utils"
           }
