+++ generators/app/templates/src/containers/Profile/components/Form/index.js
@@ -1,0 +1,3 @@
+import Form from './form.component';
+
+export default Form;
