+++ generators/app/templates/src/containers/Profile/components/Image/index.js
@@ -1,0 +1,3 @@
+import { Image } from './image.component';
+
+export default Image;
