+++ generators/app/templates/src/containers/Profile/components/index.js
@@ -1,0 +1,4 @@
+import Form from './Form';
+import Image from './Image';
+
+export { Form, Image };
