+++ generators/app/templates/src/utils/context.js
@@ -3,3 +3,4 @@ export const expandedProperty = (context: Object, property: String) => {
 
   return `${context}:${suffix}`;
 }
+
